Tab stop order added to some forms
The order in which fields jumped to by pressing the tab key followed the order in the program code, which did not always correspond to the visual position in the form. As a result, the cursor "jumped" partially within a form.

This commit adds a tab stop order for those forms concerned, so that the input cursor follows the sequence in the form.